requirements.txt
Flask==1.1.2
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.6.0.66
Pillow==9.2.0
tensorflow==2.9.1
